added crud and test, update bug

+++ services/crud.sql.ts
@@ -17,16 +17,34 @@ export async function findRecord(model: typeof Model, query: any) {
 }
 
 export async function addRecord(model: typeof Model, query: any) {
-  const data = await model.create([query]);
+  const data = await model.create(query);
   return data;
 }
 
-export async function updateRecord(model: typeof Model, query: any) {
-  const data = await model.where('id', query.id).update(query);
+export async function updateRecord(model: typeof Model, query: any, newData: any) {
+  let data;
+  if (query.id) {
+    const record = await model.find(query.id);
+    if (!record) {
+      return false
+    }
+    data = await model.where(query).update(newData);
+  } else {
+    data = await model.where(query).update(newData);
+  }
   return data;
 }
 
 export async function deleteRecord(model: typeof Model, query: any) {
-  const data = await model.deleteById(query.id);
+  let data;
+  if (query.id) {
+    const record = await model.find(query.id);
+    if (!record) {
+      return false
+    }
+    data = await model.deleteById(query.id);
+  } else {
+    data = await model.where(query).delete();
+  }
   return data;
 }
